style(browsing): split file
