+++ backend/matcha/db/db.py
@@ -149,29 +149,6 @@ def db_get_user_images(id_user):
     return fetchall_to_array(filenames)
 
 
-def db_browsing_gender_sexualorientation(id_user, search):
-    return db_fetchall(
-        """
-        SELECT id, username, firstname, lastname, gender, sexual_orientation, age, fame_rating
-        FROM users
-        WHERE sexual_orientation = %s
-          AND gender = %s
-          AND age BETWEEN %s AND %s
-          AND fame_rating BETWEEN %s AND %s
-          AND id != %s
-            """,
-        (
-            search["sexual_orientation"],
-            search["gender"],
-            search["min_age"],
-            search["max_age"],
-            search["min_fame"],
-            search["max_fame"],
-            id_user,
-        ),
-    )
-
-
 def db_get_iduser_per_username(username):
     return db_fetchone(
         "SELECT id FROM users WHERE username = %s",
