+++ backend/matcha/services/browsing.py
@@ -2,9 +2,10 @@ from flask import Blueprint, request, jsonify
 
 from matcha.db.db import (
     db_get_user_per_id,
-    db_browsing_gender_sexualorientation,
     db_get_url_profile,
 )
+
+from matcha.db.browsing import db_browsing_gender_sexualorientation
 
 from matcha.app_utils import check_request_json, flaskprint
